Convert mx waveform to floats as early as possible in the chain
This way we're sure that internally we always deal with [0;1] float samples. the [0;1024] int range is used only at the rust sdk boundary.

+++ features/messages/impl/src/main/kotlin/io/element/android/features/messages/impl/timeline/model/event/TimelineItemVoiceContent.kt
@@ -27,7 +27,7 @@ data class TimelineItemVoiceContent(
     val duration: Duration,
     val mediaSource: MediaSource,
     val mimeType: String,
-    val waveform: ImmutableList<Int>,
+    val waveform: ImmutableList<Float>,
 ) : TimelineItemEventContent {
     override val type: String = "TimelineItemAudioContent"
 }
